# openapi.yaml
openapi: 3.1.0
info:
  title: УчусьИИ
  description: 'API ИИ-наставника «УчусьИИ»: вход через MAX, задачи учеников, диалоги
    с наставником, аналитика педагога, дайджесты родителя, согласия и персональные
    данные (152-ФЗ). Маршруты /dev/* существуют только в режиме локальной проверки
    (DEV_LOGIN_ENABLED).'
  version: 0.1.0
paths:
...
  /api/v1/me:
    get:
      tags:
      - Кабинет
      summary: Профиль
      operationId: me_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - Кабинет
      summary: Прекращение обработки и удаление данных
      operationId: erase_me_api_v1_me_delete
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Erase Me Api V1 Me Delete
      security:
      - HTTPBearer: []
    patch:
      tags:
      - Кабинет
      summary: Обновить профиль
      operationId: update_me_api_v1_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me/phone:
    patch:
      tags:
      - Кабинет
      summary: Сохранить номер телефона из MAX
      operationId: share_phone_api_v1_me_phone_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneShareRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me/personal-data:
    get:
      tags:
      - Кабинет
      summary: Выписка об обрабатываемых данных
      operationId: personal_data_report_api_v1_me_personal_data_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalDataReport'
      security:
      - HTTPBearer: []
...
  /api/v1/legal/{code}:
    get:
      tags:
      - Персональные данные
      summary: Текст документа
      operationId: get_document_api_v1_legal__code__get
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/consents:
    get:
      tags:
      - Персональные данные
      summary: Состояние согласий
      operationId: my_consents_api_v1_consents_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ConsentState'
                type: array
                title: Response My Consents Api V1 Consents Get
      security:
      - HTTPBearer: []
  /api/v1/consents/{purpose}:
    post:
      tags:
      - Персональные данные
      summary: Предоставить согласие
      operationId: grant_consent_api_v1_consents__purpose__post
      security:
      - HTTPBearer: []
      parameters:
      - name: purpose
        in: path
        required: true
        schema:
          type: string
          title: Purpose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentDecision'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsentState'
                title: Response Grant Consent Api V1 Consents  Purpose  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Персональные данные
      summary: Отозвать согласие
      operationId: revoke_consent_api_v1_consents__purpose__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: purpose
        in: path
        required: true
        schema:
          type: string
          title: Purpose
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsentState'
                title: Response Revoke Consent Api V1 Consents  Purpose  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personal-data/erasure-request:
    post:
      tags:
      - Персональные данные
      summary: Запрос на удаление персональных данных
      operationId: erasure_request_api_v1_personal_data_erasure_request_post
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Erasure Request Api V1 Personal Data Erasure Request
                  Post
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ConsentDecision:
      properties:
        accept:
          type: boolean
          title: Accept
          default: true
      type: object
      title: ConsentDecision
    ConsentState:
      properties:
        purpose:
          type: string
          title: Purpose
        granted:
          type: boolean
          title: Granted
        policy_code:
          type: string
          title: Policy Code
        policy_version:
          type: string
          title: Policy Version
        granted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Granted At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
      type: object
      required:
      - purpose
      - granted
      - policy_code
      - policy_version
      title: ConsentState
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PersonalDataReport:
      properties:
        profile:
          $ref: '#/components/schemas/UserView'
        consents:
          items:
            $ref: '#/components/schemas/ConsentState'
          type: array
          title: Consents
        tasks_total:
          type: integer
          title: Tasks Total
        notifications_total:
          type: integer
          title: Notifications Total
        audit_entries:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Audit Entries
      type: object
      required:
      - profile
      - consents
      - tasks_total
      - notifications_total
      - audit_entries
      title: PersonalDataReport
    PhoneShareRequest:
      properties:
        phone:
          type: string
          maxLength: 20
          minLength: 10
          title: Phone
        auth_date:
          type: string
          maxLength: 20
          minLength: 8
          title: Auth Date
        hash:
          type: string
          maxLength: 128
          minLength: 32
          title: Hash
      type: object
      required:
      - phone
      - auth_date
      - hash
      title: PhoneShareRequest
    PolicyView:
      properties:
        code:
          type: string
          title: Code
        version:
          type: string
          title: Version
        title:
          type: string
          title: Title
        body:
          type: string
          title: Body
      type: object
      required:
      - code
      - version
      - title
      - body
      title: PolicyView
    ProfileUpdate:
      properties:
        email:
          anyOf:
          - type: string
            maxLength: 254
          - type: 'null'
          title: Email
      type: object
      title: ProfileUpdate
...
    UserView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        role:
          type: string
          title: Role
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        locale:
          type: string
          title: Locale
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
        is_active:
          type: boolean
          title: Is Active
        has_phone:
          type: boolean
          title: Has Phone
          default: false
        has_email:
          type: boolean
          title: Has Email
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - role
      - locale
      - is_active
      - created_at
      title: UserView
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      description: JWT доступа мини-приложения
      scheme: bearer